+++ src/app/pages/webinars/webinar-cards/paginated-webinar-grid.scss
@@ -1,0 +1,10 @@
+@import 'pattern-library/core/pattern-library/headers';
+
+.webinars.page {
+    .paginated-webinar-grid {
+        display: flex;
+        flex-direction: column;
+        gap: 2rem;
+        width: 100%;
+    }
+}
