+++ src/app/pages/webinars/webinar-cards/webinar-grid-section.tsx
@@ -10,6 +10,7 @@ type Args = {
     filter?: WebinarFilter;
     limit?: number;
 };
+
 export default function LabeledWebinarGrid({
     heading,
     webinars,
