+++ src/app/pages/webinars/webinar-cards/webinar-grid.scss
@@ -30,7 +30,7 @@
 
         &.past {
             gap: $normal-margin;
-            grid-template-rows: auto 1fr auto;
+            grid-template-rows: auto auto 1fr auto;
         }
 
         &.upcoming {
